requirements.txt
torch>=2.2
torchvision>=0.17
opencv-python>=4.8
numpy>=1.24
PyYAML>=6.0
tqdm>=4.66
albumentations>=1.3.1
